server/requirements.txt
fastapi
uvicorn[standard]
Pillow
httpx
python-multipart
colorama
numpy